+++ src/TestUtils/TestLdapConnectionProvider.php
@@ -22,15 +22,18 @@ class TestLdapConnectionProvider extends LdapConnectionProvider
         Configuration::CONNECTIONS_ATTRIBUTE => [
             self::DEFAULT_CONNECTION_IDENTIFIER => [
                 Configuration::LDAP_HOST_ATTRIBUTE => 'ldap.com',
-                Configuration::LDAP_BASE_DN_ATTRIBUTE => 'dc=example,dc=com',
+                Configuration::LDAP_BASE_DN_ATTRIBUTE => self::BASE_DN,
                 Configuration::LDAP_USERNAME_ATTRIBUTE => 'user',
                 Configuration::LDAP_PASSWORD_ATTRIBUTE => 'secret',
                 Configuration::LDAP_ENCRYPTION_ATTRIBUTE => 'start_tls',
-                Configuration::LDAP_OBJECT_CLASS_ATTRIBUTE => 'person',
+                Configuration::LDAP_OBJECT_CLASS_ATTRIBUTE => self::OBJECT_CLASS,
                 Configuration::LDAP_NUM_RESULT_ITEMS_WILL_SORT_LIMIT_ATTRIBUTE => 3,
             ],
         ],
     ];
+
+    private const BASE_DN = 'dc=example,dc=com';
+    private const OBJECT_CLASS = 'person';
 
     public static function create(): TestLdapConnectionProvider
     {
@@ -42,12 +45,34 @@ class TestLdapConnectionProvider extends LdapConnectionProvider
         return $provider;
     }
 
+    /**
+     * values used in exceptions need to be converted to the internal query representation.
+     */
+    public static function toExpectedValue(string $input): string
+    {
+        $output = '';
+        foreach (str_split($input) as $char) {
+            $output .= '\\'.dechex(ord($char));
+        }
+
+        return $output;
+    }
+
+    public static function getObjectClassCriteria(): string
+    {
+        return 'objectClass='.self::toExpectedValue(self::OBJECT_CLASS);
+    }
+
     public function useInApiTest(\Symfony\Component\DependencyInjection\Container $container): void
     {
         $container->set(LdapConnectionProvider::class, $this);
     }
 
-    public function mockResults(array $results = [], ?string $expectCn = null, string $expectObjectClass = 'person',
+    /**
+     * @param callable(string): bool|null $isQueryAsExpected
+     */
+    public function mockResults(array $results = [],
+        ?string $expectCn = null, ?callable $isQueryAsExpected = null,
         string $mockConnectionIdentifier = self::DEFAULT_CONNECTION_IDENTIFIER): void
     {
         $ldapExpectation = LdapFake::operation('search')
@@ -55,10 +80,15 @@ class TestLdapConnectionProvider extends LdapConnectionProvider
             ->andReturn($results);
 
         if ($expectCn !== null) {
-            $expectCnOctal = self::convertStringToOctal($expectCn);
-            $expectObjectClassOctal = self::convertStringToOctal($expectObjectClass);
-            $expectQuery = "(&(objectClass=$expectObjectClassOctal)(cn=$expectCnOctal))";
-            $ldapExpectation->with('dc=example,dc=com', $expectQuery);
+            $isQueryAsExpected = function (string $query) use ($expectCn): bool {
+                $expectCnHex = self::toExpectedValue($expectCn);
+                $objectClassCriteria = self::getObjectClassCriteria();
+
+                return $query === "(&($objectClassCriteria)(cn=$expectCnHex))";
+            };
+        }
+        if ($isQueryAsExpected !== null) {
+            $ldapExpectation->with(self::BASE_DN, $isQueryAsExpected);
         }
 
         $this->getFakeLdap($mockConnectionIdentifier)
@@ -109,15 +139,5 @@ class TestLdapConnectionProvider extends LdapConnectionProvider
         assert($ldap instanceof LdapFake);
 
         return $ldap;
-    }
-
-    private static function convertStringToOctal(string $input): string
-    {
-        $output = '';
-        foreach (str_split($input) as $char) {
-            $output .= '\\'.dechex(ord($char));
-        }
-
-        return $output;
     }
 }